Split components from llm modal
